meglio

+++ printf.h
@@ -6,7 +6,7 @@
 /*   By: luevange <luevange@student.42roma.it>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/04 00:26:28 by luevange          #+#    #+#             */
-/*   Updated: 2025/01/12 17:21:07 by luevange         ###   ########.fr       */
+/*   Updated: 2025/01/12 23:18:02 by luevange         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,6 +27,6 @@ int	ft_put_str(char *s);
 int	ft_put_pntr(unsigned long pnt, const char *s);
 int	ft_put_nbr(int n);
 int	ft_put_hex_nmb(unsigned int n, const char *s);
-int	ft_put_uns_nmb(unsigned int n);
+int	ft_put_uns_nmb(unsigned long long n);
 
 #endif
